Move ByteTests
